[bug/12] Move src/decode-json.ts → src/index.ts
