+++ .prettierignore
@@ -9,3 +9,4 @@ yarn.lock
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+.nyc_output
