+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "decode-json",
-  "version": "1.0.0",
+  "version": "1.0.2",
   "description": "Convert unknown JSON values to TypeScript safe structures",
   "license": "BSD-3-Clause",
   "homepage": "https://github.com/owanturist/decode-json",
@@ -13,26 +13,17 @@
     "elm"
   ],
   "author": "Anton Ovechkin <owanturist@gmail.com> (https://github.com/owanturist)",
-  "types": "types/decode-json.d.ts",
-  "source": "src/decode-json.js",
-  "main": "lib/decode-json.js",
-  "unpkg": "dist/decode-json.min.js",
-  "module": "es/decode-json.js",
+  "types": "types",
+  "source": "src",
+  "main": "lib",
+  "unpkg": "dist",
+  "module": "es",
   "exports": {
-    ".": {
-      "import": "./es/decode-json.js",
-      "require": "./lib/decode-json.js",
-      "node": "./lib/decode-json.js",
-      "browser": "./es/decode-json.mjs",
-      "default": "./lib/decode-json.js"
-    },
-    "./error-to-human-readable": {
-      "import": "./es/error-to-human-readable.js",
-      "require": "./lib/error-to-human-readable.js",
-      "node": "./lib/error-to-human-readable.js",
-      "browser": "./es/error-to-human-readable.mjs",
-      "default": "./lib/error-to-human-readable.js"
-    }
+    "import": "./es",
+    "require": "./lib",
+    "node": "./lib",
+    "browser": "./es",
+    "default": "./lib"
   },
   "files": [
     "dist",
@@ -45,6 +36,8 @@
     "clean": "rm -rf ./dist ./es ./lib ./types",
     "prebuild": "npm run clean",
     "build": "rollup --config",
+    "prepublish": "npm run build",
+    "publish": "git push --tags && npm publish",
     "test": "ava",
     "test:watch": "npm test -- --watch",
     "test:coverage": "nyc ava",
@@ -90,5 +83,6 @@
     "ts-node": "^9.1.1",
     "ts-transformer-minify-privates": "^0.3.0",
     "typescript": "^4.1.3"
-  }
+  },
+  "sideEffects": false
 }
