+++ src/error-to-human-readable.ts
@@ -1,4 +1,4 @@
-import { DecodeJsonError, DecodeError } from './decode-json'
+import { DecodeJsonError, DecodeError } from '.'
 
 const wrapFieldName = (name: string): string => {
   if (/^[a-z$_][0-9a-z$_]*$/i.test(name)) {
