+++ tests/chain.test.ts
@@ -1,6 +1,6 @@
 import test from 'ava'
 
-import Decode from '../src/decode-json'
+import Decode from '../src'
 import { RequiredField } from './error'
 
 test('Decoder.chain()', t => {
