+++ tests/list.test.ts
@@ -2,7 +2,7 @@
 
 import test from 'ava'
 
-import Decode from '../src/decode-json'
+import Decode from '../src'
 import { Optional, InField, AtIndex, ExpectArray, ExpectInt } from './error'
 
 test('Decode.list()', t => {
