+++ tests/one-of.test.ts
@@ -2,7 +2,7 @@
 
 import test from 'ava'
 
-import Decode, { Decoder } from '../src/decode-json'
+import Decode, { Decoder } from '../src'
 import {
   OneOf,
   InField,
